+++ tsdoc/src/index.ts
@@ -14,7 +14,7 @@ export { TextRange, ITextLocation } from './parser/TextRange';
 export { Token, TokenKind } from './parser/Token';
 export { TokenSequence, ITokenSequenceParameters } from './parser/TokenSequence';
 export { TSDocParser } from './parser/TSDocParser';
-export { TSDocParserConfiguration } from './parser/TSDocParserConfiguration';
+export { TSDocParserConfiguration, TSDocParserValidationConfiguration } from './parser/TSDocParserConfiguration';
 export {
   ITSDocTagDefinitionParameters,
   TSDocTagSyntaxKind,