+++ tsdoc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log - @microsoft/tsdoc
 
+## 0.8.0
+- Introduce a distinction between "defined" tags (i.e. recognized) versus "supported" tags (i.e. implemented by the tool)
+- The parser optionally reports usage of undefined tags
+- The parser optionally reports usage of unsupported tags
+- The parser reports usage of inline/block syntax that is inconsistent with the tag definition
+- Code spans are no allowed to be adjacent to other text, but must contain at least one character
+- An `@deprecated` block must contain a deprecation message
+- If `@inheritDoc` is used, then the summary section must be empty, and there must not be an `@remarks` block
+
 ## 0.7.0
 - Add support for `@defaultValue` tag
 - Add support for `@typeParam` tag